requirements.txt
python-telegram-bot==20.0
nest-asyncio==1.5.8
python-docx==0.8.11